+++ old/app/js/AppDispatcher.js
@@ -1,6 +1,0 @@
-/**
- * Created by Ben on 29/05/2015.
- */
-var Dispatcher = require('flux').Dispatcher;
-
-module.exports = new Dispatcher();