+++ src/components/App/App.js
@@ -10,28 +10,48 @@
 import React, { Component, PropTypes } from 'react';
 import emptyFunction from 'fbjs/lib/emptyFunction';
 import s from './App.scss';
-import Header from '../Header';
-import Feedback from '../Feedback';
-import Footer from '../Footer';
 
+import PlayerStore from '../../stores/PlayerStore';
+import Socket from '../../api/Socket';
+import {getPlaylist, nowPlaying} from '../../actions/PlayerActionCreators';
+
+import Header from '../Header';
+import NowPlaying from '../NowPlaying';
+import Playlist from '../Playlist';
+import PlayerInfo from '../PlayerInfo';
+import Search from '../Search';
+
+function getStateFromStores() {
+  return {
+    playlist: PlayerStore.getPlaylist(true),
+    nowPlaying: PlayerStore.nowPlaying(),
+    isMaster: PlayerStore.isMaster(),
+    playerState: PlayerStore.getPlayerState()
+  }
+}
 class App extends Component {
+
+  constructor(props) {
+    super(props);
+    this.state = getStateFromStores();
+  }
 
   static propTypes = {
     context: PropTypes.shape({
       insertCss: PropTypes.func,
       onSetTitle: PropTypes.func,
       onSetMeta: PropTypes.func,
-      onPageNotFound: PropTypes.func,
+      onPageNotFound: PropTypes.func
     }),
-    children: PropTypes.element.isRequired,
-    error: PropTypes.object,
+    //children: PropTypes.element.isRequired,
+    error: PropTypes.object
   };
 
   static childContextTypes = {
     insertCss: PropTypes.func.isRequired,
     onSetTitle: PropTypes.func.isRequired,
     onSetMeta: PropTypes.func.isRequired,
-    onPageNotFound: PropTypes.func.isRequired,
+    onPageNotFound: PropTypes.func.isRequired
   };
 
   getChildContext() {
@@ -40,27 +60,45 @@ class App extends Component {
       insertCss: context.insertCss || emptyFunction,
       onSetTitle: context.onSetTitle || emptyFunction,
       onSetMeta: context.onSetMeta || emptyFunction,
-      onPageNotFound: context.onPageNotFound || emptyFunction,
+      onPageNotFound: context.onPageNotFound || emptyFunction
     };
+  }
+
+  onStoreChange() {
+    this.setState(getStateFromStores());
   }
 
   componentWillMount() {
     this.removeCss = this.props.context.insertCss(s);
   }
 
+  componentDidMount() {
+    PlayerStore.addChangeListener(this.onStoreChange.bind(this));
+    Socket.init();
+    require('../../actions/SocketActionCreators');
+    getPlaylist(true);
+    nowPlaying();
+  }
+
   componentWillUnmount() {
     this.removeCss();
+    PlayerStore.removeChangeListener(this.onStoreChange.bind(this));
   }
 
   render() {
     return !this.props.error ? (
       <div>
         <Header />
-        {this.props.children}
-        <Feedback />
-        <Footer />
+        <NowPlaying song={this.state.nowPlaying}
+                    isMaster={this.state.isMaster}
+                    playerState={this.state.playerState} />
+        <Search />
+        <div className={s.mainContent}>
+          <Playlist playlist={this.state.playlist} />
+          <PlayerInfo song={this.state.nowPlaying} />
+        </div>
       </div>
-    ) : this.props.children;
+    ) : <noscript />;
   }
 
 }