+++ src/constants/ActionTypes.js
@@ -10,5 +10,12 @@
 import keyMirror from 'fbjs/lib/keyMirror';
 
 export default keyMirror({
-
+  LOAD_SONG: null,
+  LOAD_PLAYLIST: null,
+  NOW_PLAYING: null,
+  PLAY: null,
+  PAUSE: null,
+  PLAYER_STATE: null,
+  SELECT_SONG: null,
+  SET_MASTER: null
 });