+++ tools/webpack.config.js
@@ -22,11 +22,11 @@ const AUTOPREFIXER_BROWSERS = [
   'Explorer >= 9',
   'iOS >= 7',
   'Opera >= 12',
-  'Safari >= 7.1',
+  'Safari >= 7.1'
 ];
 const GLOBALS = {
   'process.env.NODE_ENV': DEBUG ? '"development"' : '"production"',
-  __DEV__: DEBUG,
+  __DEV__: DEBUG
 };
 
 //
@@ -37,7 +37,7 @@ const GLOBALS = {
 const config = {
   output: {
     publicPath: '/',
-    sourcePrefix: '  ',
+    sourcePrefix: '  '
   },
 
   cache: DEBUG,
@@ -52,15 +52,15 @@ const config = {
     chunks: VERBOSE,
     chunkModules: VERBOSE,
     cached: VERBOSE,
-    cachedAssets: VERBOSE,
+    cachedAssets: VERBOSE
   },
 
   plugins: [
-    new webpack.optimize.OccurenceOrderPlugin(),
+    new webpack.optimize.OccurenceOrderPlugin()
   ],
 
   resolve: {
-    extensions: ['', '.webpack.js', '.web.js', '.js', '.jsx', '.json'],
+    extensions: ['', '.webpack.js', '.web.js', '.js', '.jsx', '.json']
   },
 
   module: {
@@ -69,40 +69,40 @@ const config = {
         test: /\.jsx?$/,
         include: [
           path.resolve(__dirname, '../node_modules/react-routing/src'),
-          path.resolve(__dirname, '../src'),
+          path.resolve(__dirname, '../src')
         ],
-        loader: 'babel-loader',
+        loader: 'babel-loader'
       }, {
         test: /\.scss$/,
         loaders: [
           'isomorphic-style-loader',
           `css-loader?${DEBUG ? 'sourceMap&' : 'minimize&'}modules&localIdentName=` +
           `${DEBUG ? '[name]_[local]_[hash:base64:3]' : '[hash:base64:4]'}`,
-          'postcss-loader?parser=postcss-scss',
-        ],
+          'postcss-loader?parser=postcss-scss'
+        ]
       }, {
         test: /\.json$/,
-        loader: 'json-loader',
+        loader: 'json-loader'
       }, {
         test: /\.txt$/,
-        loader: 'raw-loader',
+        loader: 'raw-loader'
       }, {
         test: /\.(png|jpg|jpeg|gif|svg|woff|woff2)$/,
-        loader: 'url-loader?limit=10000',
+        loader: 'url-loader?limit=10000'
       }, {
         test: /\.(eot|ttf|wav|mp3)$/,
-        loader: 'file-loader',
-      },
-    ],
+        loader: 'file-loader'
+      }
+    ]
   },
 
   postcss: function plugins(bundler) {
     return [
       require('postcss-import')({ addDependencyTo: bundler }),
       require('precss')(),
-      require('autoprefixer')({ browsers: AUTOPREFIXER_BROWSERS }),
+      require('autoprefixer')({ browsers: AUTOPREFIXER_BROWSERS })
     ];
-  },
+  }
 };
 
 //
@@ -113,7 +113,7 @@ const clientConfig = extend(true, {}, config, {
   entry: './src/client.js',
   output: {
     path: path.join(__dirname, '../build/public'),
-    filename: DEBUG ? '[name].js?[hash]' : '[name].[hash].js',
+    filename: DEBUG ? '[name].js?[hash]' : '[name].[hash].js'
   },
 
   // Choose a developer tool to enhance debugging
@@ -124,7 +124,7 @@ const clientConfig = extend(true, {}, config, {
     new AssetsPlugin({
       path: path.join(__dirname, '../build'),
       filename: 'assets.js',
-      processOutput: x => `module.exports = ${JSON.stringify(x)};`,
+      processOutput: x => `module.exports = ${JSON.stringify(x)};`
     }),
     ...(!DEBUG ? [
       new webpack.optimize.DedupePlugin(),
@@ -134,12 +134,12 @@ const clientConfig = extend(true, {}, config, {
           screw_ie8: true,
 
           // jscs:enable requireCamelCaseOrUpperCaseIdentifiers
-          warnings: VERBOSE,
-        },
+          warnings: VERBOSE
+        }
       }),
-      new webpack.optimize.AggressiveMergingPlugin(),
-    ] : []),
-  ],
+      new webpack.optimize.AggressiveMergingPlugin()
+    ] : [])
+  ]
 });
 
 //
@@ -151,7 +151,7 @@ const serverConfig = extend(true, {}, config, {
   output: {
     path: './build',
     filename: 'server.js',
-    libraryTarget: 'commonjs2',
+    libraryTarget: 'commonjs2'
   },
   target: 'node',
   externals: [
@@ -162,7 +162,7 @@ const serverConfig = extend(true, {}, config, {
         !request.match(/^react-routing/) &&
         !context.match(/[\\/]react-routing/);
       cb(null, Boolean(isExternal));
-    },
+    }
   ],
   node: {
     console: false,
@@ -170,14 +170,14 @@ const serverConfig = extend(true, {}, config, {
     process: false,
     Buffer: false,
     __filename: false,
-    __dirname: false,
+    __dirname: false
   },
   devtool: 'source-map',
   plugins: [
     new webpack.DefinePlugin({ ...GLOBALS, 'process.env.BROWSER': false }),
     new webpack.BannerPlugin('require("source-map-support").install();',
-      { raw: true, entryOnly: false }),
-  ],
+      { raw: true, entryOnly: false })
+  ]
 });
 
 export default [clientConfig, serverConfig];